+++ net-inserter-async/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod component;
-pub mod utils;
-pub mod config;