+++ net-reporter/src/config.rs
@@ -11,6 +11,12 @@ use std::env;
 use net_config::NetConfig;
 
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
+pub struct VerifyToken {
+    pub token: bool,
+    pub default_token: String,
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
 pub struct Server {
     pub host_name: String,
     pub port: String,
@@ -39,6 +45,7 @@ pub struct FusionAuthApiKey {
 
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq, NetConfig)]
 pub struct Config {
+    pub verify_token: VerifyToken,
     pub server: Server,
     pub connection_url: ConnectionUrl,
     pub max_connection_size: MaxConnectionSize,