+++ net-reporter/src/query/charts/network_graph/request/requester.rs
@@ -86,7 +86,7 @@ impl Requester for NetworkGraphRequester {
 
         let executed_query_response = Self::execute_query(
             connection_pool,
-            Some(jwt.get_tenant_id()),
+            jwt.get_tenant_id(),
             request_start_date,
             request_end_date,
             filters,