+++ net-reporter/src/query/manager/query_manager.rs
@@ -3,6 +3,7 @@ use std::error::Error;
 use std::sync::Arc;
 
 use net_token_verifier::fusion_auth::fusion_auth_verifier;
+use net_token_verifier::fusion_auth::jwt_token::JwtBuilder;
 use net_token_verifier::verifier::Verifier;
 use sqlx::Pool;
 use sqlx::Postgres;
@@ -45,14 +46,19 @@ impl QueryManager {
         if enveloped_request.get_jwt_token().is_err() {
             return Err("error: jwt token is required".into());
         }
-        let jwt = fusion_auth_verifier::FusionAuthVerifier::new(
-            &config.fusion_auth_server_addres.addr,
-            Some(config.fusion_auth_api_key.key.clone()))
-            .verify_token(enveloped_request.get_jwt_token().unwrap()).await;
-
-        let jwt = match jwt {
-            Ok(jwt) => jwt,
-            Err(e) => return Err(format!("error: {:?}", e).into())
+        let jwt = if config.verify_token.token {
+            let jwt = fusion_auth_verifier::FusionAuthVerifier::new(
+                &config.fusion_auth_server_addres.addr,
+                Some(config.fusion_auth_api_key.key.clone()))
+                .verify_token(enveloped_request.get_jwt_token().unwrap()).await;
+            match jwt {
+                Ok(jwt) => jwt,
+                Err(e) => return Err(format!("error: {:?}", e).into())
+            }
+        } else {
+            JwtBuilder::default()
+                .with_tenant_id(config.verify_token.default_token.clone())
+                .build()
         };
 
         let requester = requester.unwrap().as_ref();